requirements.txt
Click~=7.0
Flask~=1.0.2
ipdb~=0.11
ipython~=7.1.1
RPi.GPIO~=0.6.4
boto3